iterator: move test iterators to it's own file and unexport it
